refactor(vendor-codex): move extractAccountId to utils module
- Move extractAccountId function from auth.ts to new utils.ts file
- Update imports in model.ts to use the new utils module
- Clean up unused AuthClaims import in auth.ts
