+++ packages/vendor-codex/src/model.ts
@@ -1,9 +1,9 @@
 import { createOpenAI } from "@ai-sdk/openai";
 import type { CreateModelOptions } from "@getpochi/common/vendor/edge";
 import { wrapLanguageModel } from "ai";
-import { extractAccountId } from "./auth";
 import { transformToCodexFormat } from "./transformers";
 import type { CodexCredentials } from "./types";
+import { extractAccountId } from "./utils";
 
 /**
  * Create headers for Codex API requests
